pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "polymarketbot"
version = "0.1.0"
description = "High-performance Polymarket copy-trading bot"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "PolyBot" }
]

dependencies = [
    # HTTP & Async
    "httpx>=0.27.0",
    "websockets>=12.0",
    "aiohttp>=3.9.0",

    # Web3 & Signing
    "web3>=6.15.0",
    "eth-account>=0.11.0",
    "py-clob-client>=0.18.0",

    # Configuration
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",

    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",

    # CLI
    "typer>=0.9.0",
    "rich>=13.7.0",

    # Utilities
    "structlog>=24.1.0",
    "tenacity>=8.2.0",
    "sortedcontainers>=2.4.0",

    # Monitoring (optional for dev)
    "prometheus-client>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
polybot = "src.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
